Just to save some time
I copied and pasted the general format of each description and labeled each HTML element. I dont know why but the txt progressively increased in size. I dont know if that if because of the copy/paste function. I forgot to comment on the last commit so i'd like to add that I feel unsure about what info. to include for content and support. I dont want to write too much but also not too little